% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_details.R
\name{download_details}
\alias{download_details}
\title{download details of artworks to local files}
\usage{
download_details(df = realism, n = 10)
}
\arguments{
\item{df}{The dataframe returned by the get_search_results() function. Default to realism}

\item{n}{The number of artworks you want to download. Default to 10}
}
\description{
This function should be used after the get_search_results() function. Users can specify how many artworks they want to download, and the function can download the images, maps and realted artworks.
}
\examples{
download_details(n = 120)
}
\keyword{download}
