% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_represents.R
\name{find_represent_artists}
\alias{find_represent_artists}
\title{Find represent artists}
\usage{
find_represent_artists(df = realism)
}
\arguments{
\item{df}{The dataframe returned by the get_search_results() function. Default to realism}
}
\value{
Dataframe
}
\description{
This function should be used after get_search_results(), it can help users to analyze the search results can return the artists who create the most artworks of the theme.
}
\examples{
find_represent_artists()
}
