% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_represents.R
\name{find_represent_dep}
\alias{find_represent_dep}
\title{Find represent departments}
\usage{
find_represent_dep(df = realism)
}
\arguments{
\item{df}{The dataframe returned by the get_search_results() function. Default to realism}
}
\value{
Dataframe and plot
}
\description{
This function should be used after get_search_results(), it can help users to analyze the search results can return the departments which have the most artworks of the theme.
}
\examples{
find_represent_dep()
}
