% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_art_age.R
\name{get_art_age}
\alias{get_art_age}
\title{Find artworks in certain age}
\usage{
get_art_age(Begin = 1600, End = 1900, class = "Paintings")
}
\arguments{
\item{Begin}{The start of the period. Defaults to 1600}

\item{End}{The end of the period. Defaults to 1900}

\item{class}{The type of artworks you want to choose. Defaults to Paintings}
}
\value{
Dataframe
}
\description{
This function allows you to search a type of artworks during a period. You can specify the type of artworks, the start and the end of the period.
}
\examples{
get_art_age(Begin = 1600, End = 1900, class = "Paintings")
get_art_age(Begin = 1620, End = 1650, class = "Sculpture")

}
\keyword{age}
