% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artworks_in_dep.R
\name{get_artworks_in_dep}
\alias{get_artworks_in_dep}
\title{Get artworks in certain departments}
\usage{
get_artworks_in_dep(
  dep_list = c("American Decorative Arts", "Arms and Armor", "Greek and Roman Art"),
  metadataDate = "2019-11-30"
)
}
\arguments{
\item{dep_list}{a list of names of departments that you want to know their ids. Defaults to c('American Decorative Arts','Arms and Armor','Greek and Roman Art').}

\item{metadataDate}{the function will return any objects with updated data after this date. Defaults to "2019-11-30".}
}
\value{
Dataframe
}
\description{
This function allows you to get details of artworks in the departments you are interested in.
}
\examples{
get_artworks_in_dep()
get_artworks_in_dep(metadataDate = "2019-12-9")
}
