% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_departments.R
\name{get_dep_ids}
\alias{get_dep_ids}
\title{Get departments ids}
\usage{
get_dep_ids(
  dep_list = c("American Decorative Arts", "Arms and Armor", "Greek and Roman Art")
)
}
\arguments{
\item{dep_list}{a list of names of departments that you want to know their ids.Defaults to c('American Decorative Arts','Arms and Armor','Greek and Roman Art').}
}
\description{
This function will return department ids separated by "|" when users input a list of department names
}
\examples{
get_dep_ids()

}
