% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_departments.R
\name{get_departments}
\alias{get_departments}
\title{Get departments ids and names}
\usage{
get_departments()
}
\value{
Dataframe
}
\description{
This function allows you to get the ids and names of departments in the MET museum.
}
\examples{
get_departments()

}
\keyword{department}
