% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_art_age.R
\name{get_distribution_country}
\alias{get_distribution_country}
\title{Find the number of artworks in country}
\usage{
get_distribution_country(df = artworks_in_age)
}
\arguments{
\item{df}{The dataframe returned by the function get_art_age(). Defaults to artworks_in_age}
}
\description{
This function ranks the country according to the number of artworks in the period you choose in get_art_age() function.
}
\examples{
get_distribution_country()

}
