% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{download map of certain artwork}
\usage{
get_map(objectIDs = 547802, path = "maps/")
}
\arguments{
\item{objectIDs}{The id of the artwork. Default to 547802}

\item{path}{the function will create a file folder with the name of path to store the map. Default to "maps/", so it will create a file folder called maps and store files.}
}
\description{
This function allows users to download the maps of artworks by inputting the objectIDs of the artwork.
}
\examples{
get_map(objectIDs = 547802)
get_map(objectIDs = 14019)

}
\keyword{map}
