% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artworks_in_dep.R
\name{get_popular_items}
\alias{get_popular_items}
\title{Get the popular artworks}
\usage{
get_popular_items(df = artworks_in_dep)
}
\arguments{
\item{df}{The dataframe returned by the function get_artworks_in_dep(). Defaults to artworks_in_dep}
}
\value{
Dataframe
}
\description{
This function should be used after calling the get_artworks_in_dep() function, and this function can return the details of popular artworks in the departments you choose.
}
\examples{
get_popular_items()
}
