% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artworks_in_dep.R
\name{get_popular_perc}
\alias{get_popular_perc}
\title{Compare the percentage of popular artworks}
\usage{
get_popular_perc(df = artworks_in_dep)
}
\arguments{
\item{df}{The dataframe returned by the function get_artworks_in_dep(). Defaults to artworks_in_dep}
}
\value{
Dataframe
}
\description{
This function should be used after calling the get_artworks_in_dep() function, and this function can calculate the percentage of popular artworks in those departments you choose in the last function.
}
\examples{
get_popular_perc()
}
