% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_related_artworks.R
\name{get_related_artworks}
\alias{get_related_artworks}
\title{get a dataframe of realted artworks about certain artwork}
\usage{
get_related_artworks(objectIDs = 547802)
}
\arguments{
\item{objectIDs}{The id of the artwork. Default to 547802}
}
\description{
This function will crawl the related artworks about certain artwork, and users can specify the artwork by objectIDs.
}
\examples{
get_related_artworks(objectIDs = 547802)

}
\keyword{relatedartworks}
