% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_art_age.R
\name{sort_time_nation}
\alias{sort_time_nation}
\title{Sort the dataframe by time and nationality}
\usage{
sort_time_nation(df = artworks_in_age)
}
\arguments{
\item{df}{The dataframe returned by the function get_art_age(). Defaults to artworks_in_age}
}
\description{
This function should be used after the function get_art_age(), and this function can sort the data by nationality and the created age.
}
\examples{
sort_time_nation()

}
